requirements.txt
absl-py==1.0.0
addict==2.4.0
ale-py==0.7.4
alembic==1.10.4
anyio==3.6.1
apsw==3.28.0.post1
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
atari-py==0.2.9
atomicwrites==1.1.5
attrs==20.3.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
Babel==2.10.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
black==23.3.0
bleach==3.2.1
blinker==1.4
bottle==0.12.25
Brlapi==0.7.0
bs4==0.0.1
cachetools==5.0.0
ccsm==0.9.14.1
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
click==8.0.4
cloudpickle==2.2.0
cmaes==0.9.1
cmake==3.26.3
cmudict==1.0.2
colorama==0.4.4
colorlog==6.7.0
command-not-found==0.3
compizconfig-python==0.9.14.1
coverage==7.2.7
cryptography==36.0.1
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cupy-cuda115==9.6.0
cv==1.0.0
cvxopt==1.2.7
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
deprecation==2.1.0
devscripts===2.20.2ubuntu2
distro==1.4.0
distro-info==0.23+ubuntu1.1
dm-env==1.6
dm-tree==0.1.6
dnspython==1.16.0
docker-pycreds==0.4.0
duplicity==0.8.12.0
eduvpn-client==4.3.1
eduvpn-common==2.0.2
entrypoints==0.3
exceptiongroup==1.1.1
f.lux indicator applet==1.2.1rc0
Farama-Notifications==0.0.4
fasteners==0.15
fastjsonschema==2.15.3
fastrlock==0.8
feedparser==5.2.1
filelock==3.6.0
flatbuffers==2.0
fonttools==4.29.1
free-mujoco-py==2.1.6
future==0.18.2
gast==0.5.3
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
glfw==1.12.0
google-api-core==2.7.1
google-api-python-client==2.39.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.55.0
gpg==1.13.1
GPUtil==1.4.0
grpcio==1.54.2
gym==0.21.0
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.6.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.20.4
idna==2.8
ifaddr==0.1.6
imageio==2.19.0
importlib-metadata==4.13.0
importlib-resources==5.4.0
iniconfig==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.12.0
jax-jumpy==1.0.0
jedi==0.17.2
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyter-packaging==0.12.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
labmaze==1.0.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
lit==16.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mdurl==0.1.2
mechanize==0.4.5
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpld3==0.5.7
mpmath==1.3.0
msgpack==0.6.2
mujoco==2.3.6
mujoco-py==2.1.2.14
mypy==1.2.0
mypy-extensions==1.0.0
nbclient==0.5.1
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==3.1
nose==1.3.7
notebook==6.1.5
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
open3d==0.15.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
optuna==3.2.0
optuna-dashboard==0.10.0
outcome==1.1.0
packaging==23.1
pandas==1.4.1
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
pathspec==0.11.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==3.5.0
plotly==5.14.1
pluggy==1.0.0
portalocker==2.4.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
pronouncing==0.2.0
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.6.0
py==1.8.1
py-cpuinfo==8.0.0
pyactr==0.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pychm==0.8.6
pycocotools==2.0.4
pycparser==2.20
pycups==1.9.73
PyDrive==1.3.1
pygame==2.1.0
Pygments==2.15.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyquaternion==0.9.9
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.7.1
pytest==7.3.1
pytest-cov==4.1.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-magic==0.4.16
python-slugify==6.1.1
pytorch-lightning==1.0.8
pytz==2021.3
pyxdg==0.26
PyYAML==5.4.1
pyzmq==20.0.0
regex==2022.3.2
reportlab==3.5.34
repoze.lru==0.7
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==13.3.5
Routes==2.4.1
rsa==4.8
ruff==0.0.264
sacrebleu==2.0.0
sacremoses==0.0.47
safetensors==0.4.4
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.1.3
Send2Trash==1.5.0
sentencepiece==0.1.96
sentry-sdk==2.13.0
seqeval==1.2.2
setproctitle==1.3.3
simplejson==3.16.0
simpy==4.0.1
sip==4.19.21
six==1.14.0
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.3.12
stable-baselines3==1.1.0
sympy==1.12
systemd-python==234
tabulate==0.8.9
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.7.0
tensorflow-model-optimization==0.7.1
tensorflow-text==2.8.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.10.2
tomotopy==0.12.2
torchsummary==1.5.1
torchvision==0.8.1+cu110
tornado==6.1
tqdm==4.63.0
traitlets==5.0.5
trio==0.20.0
trio-websocket==0.9.2
triton==3.0.0
typeguard==2.13.3
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
uritemplate==4.1.1
urllib3==1.26.19
usb-creator==0.3.7
wadllib==1.3.3
wandb==0.17.7
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.5
websocket-client==1.3.2
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.13.3
wsproto==1.1.0
xkit==0.0.0
zeroconf==0.24.4
zipp==3.7.0
